add socket function support

+++ tests/library/Rediska/TestCase.php
@@ -14,11 +14,14 @@ class Rediska_TestCase extends PHPUnit_Framework_TestCase
 
     protected function tearDown()
     {
-        $this->rediska->flushDb(true);
-        foreach($this->rediska->getConnections() as $connection) {
-            $connection->disconnect();
+        if ($this->rediska !== null) {
+            $this->rediska->flushDb(true);
+            foreach($this->rediska->getConnections() as $connection) {
+                $connection->disconnect();
+            }
+
+            $this->rediska = null;
         }
-        $this->rediska = null;
     }
 
     protected function _addSecondServerOrSkipTest()
